3차 피드백 적용
